refactor: update model logo imports and remove deprecated model configuration files
