+++ src/renderer/src/config/models/logo.ts
